ty a simpler formatter
