+++ org.rebecalang.afra.ideplugin/src/org/rebecalang/afra/ideplugin/editors/formatter/FixedPropertyFormatter.java
@@ -1,0 +1,96 @@
+package org.rebecalang.afra.ideplugin.editors.formatter;
+
+import org.eclipse.jface.text.BadLocationException;
+import org.eclipse.jface.text.IDocument;
+import org.eclipse.jface.text.IRegion;
+
+/**
+ * Fixed formatter for Property (.property) files that correctly handles comments
+ */
+public class FixedPropertyFormatter implements IAfraFormatter {
+    
+    private static final String INDENT = "\t";
+    private static final String NEW_LINE = System.getProperty("line.separator");
+    
+    @Override
+    public String format(IDocument document) {
+        try {
+            String content = document.get();
+            return formatContent(content);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return document.get();
+        }
+    }
+    
+    @Override
+    public String format(IDocument document, IRegion region) {
+        try {
+            String content = document.get(region.getOffset(), region.getLength());
+            return formatContent(content);
+        } catch (BadLocationException e) {
+            e.printStackTrace();
+            return document.get();
+        }
+    }
+    
+    @Override
+    public String getIndentString() {
+        return INDENT;
+    }
+    
+    private String formatContent(String content) {
+        if (content == null || content.trim().isEmpty()) {
+            return content;
+        }
+        
+        // Normalize line endings and remove trailing spaces
+        content = content.replaceAll("\r\n", "\n").replaceAll("\r", "\n");
+        content = content.replaceAll("[ \t]+\n", "\n");
+        
+        String[] lines = content.split("\n");
+        StringBuilder result = new StringBuilder();
+        int braceLevel = 0;
+        
+        for (String line : lines) {
+            String trimmed = line.trim();
+            
+            // Handle empty lines
+            if (trimmed.isEmpty()) {
+                result.append(NEW_LINE);
+                continue;
+            }
+            
+            // Calculate indentation level for this line
+            int currentIndent = braceLevel;
+            
+            // Closing braces decrease the indentation level for themselves
+            if (trimmed.startsWith("}")) {
+                currentIndent = Math.max(0, braceLevel - 1);
+                braceLevel = currentIndent; // Update for next lines
+            }
+            
+            // Apply indentation
+            for (int i = 0; i < currentIndent; i++) {
+                result.append(INDENT);
+            }
+            
+            // Add the line content
+            result.append(trimmed);
+            result.append(NEW_LINE);
+            
+            // Opening braces increase indentation level for subsequent lines
+            // BUT only for non-comment lines
+            if (trimmed.endsWith("{") && !isComment(trimmed)) {
+                braceLevel++;
+            }
+        }
+        
+        return result.toString().replaceAll("\n+$", "\n");
+    }
+    
+    private boolean isComment(String trimmed) {
+        return trimmed.startsWith("//") || trimmed.startsWith("/*") || 
+               trimmed.startsWith("*") || trimmed.endsWith("*/");
+    }
+}
